+++ src/main/java/com/ivietech/demo/dao/ProductDetailRepository.java
@@ -5,7 +5,6 @@
  */
 package com.ivietech.demo.dao;
 
-import com.ivietech.demo.entity.Platforms;
 import com.ivietech.demo.entity.ProductDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -15,6 +14,6 @@ import org.springframework.stereotype.Repository;
  * @author HoangMinh
  */
 @Repository
-public interface ProductDetailRepository extends JpaRepository<ProductDetail, Integer> {
+public interface ProductDetailRepository extends JpaRepository<ProductDetail, Long> {
     
 }
